+++ src/app/api/hathi/route.ts
@@ -1,20 +1,22 @@
 import { NextResponse } from "next/server";
 import { matchRegex, pause } from "@/app/helper/dataDownloader";
+import { writeFileSync } from "fs";
 
-import fs from "fs";
-
-const BASE_URL = "http://creatingdata.us/data/scatter/hathi/tiles/";
+// inspired by : https://github.com/ColinEberhardt/d3fc-webgl-hathi-explorer
+const BASE_URL = "http://creatingdata.us/data/scatter/hathi/tiles";
 
 const getData = async (url: string) => {
   let data = "";
-  const res = await fetch(url);
-  data += res.json;
+  const response = await fetch(url);
+  const chunk = await response.text();
+  data += chunk;
   return data;
 };
 
 export async function GET() {
   const tileFolders = await matchRegex(`${BASE_URL}/`, /href="([0-9]*)\/"/g);
-  //   console.log(tileFolders);
+  console.log("tileFolders", tileFolders);
+
   for (let j = 5; j < tileFolders.length; j++) {
     const tileFolder = tileFolders[j];
 
@@ -22,6 +24,7 @@ export async function GET() {
       `${BASE_URL}/${tileFolder}/`,
       /href="([0-9]*)\/"/g
     );
+
     for (let i = 0; i < subFolders.length; i++) {
       const subFolder = subFolders[i];
 
@@ -37,8 +40,9 @@ export async function GET() {
         await pause(500);
 
         const fileData = await getData(fileUrl);
-        console.log(fileData);
-        fs.writeFileSync(`data/${tileFolder}-${subFolder}-${file}`, fileData);
+        // const dirPath = path.join(__dirname, "/data");
+        // console.log(dirPath);
+        writeFileSync(`src/data/${tileFolder}-${subFolder}-${file}`, fileData);
       }
     }
   }
