+++ src/app/components/WebGLChart.tsx
@@ -1,0 +1,5 @@
+const WebGLChart = () => {
+  return <></>;
+};
+
+export default WebGLChart;
